requirements.txt
fastapi
uvicorn[standard]
pandas
pydantic
python-multipart
python-dotenv
httpx
numpy
motor

